separating repositories to attend specific useCases
