+++ app/src/main/java/com/airongomes/eventsapi/di/EventModule.kt
@@ -2,6 +2,8 @@ package com.airongomes.eventsapi.di
 
 import com.airongomes.eventsapi.domain.remote.api.EventApi
 import com.airongomes.eventsapi.domain.remote.instantiateApi
+import com.airongomes.eventsapi.domain.repository.CheckInRepository
+import com.airongomes.eventsapi.domain.repository.CheckInRepositoryImpl
 import com.airongomes.eventsapi.domain.repository.EventRepository
 import com.airongomes.eventsapi.domain.repository.EventRepositoryImpl
 import com.airongomes.eventsapi.domain.usecase.CheckInUseCase
@@ -16,6 +18,7 @@ import org.koin.dsl.module
 val eventModule = module {
     single { instantiateApi(EventApi::class.java) }
     single<EventRepository> { EventRepositoryImpl(get()) }
+    single<CheckInRepository> { CheckInRepositoryImpl(get()) }
     factory<FetchEventsUseCase> { FetchEventsUseCaseImpl(get()) }
     factory<CheckInUseCase> { CheckInUseCaseImpl(get()) }
