+++ app/src/main/java/com/airongomes/eventsapi/domain/usecase/CheckInUseCase.kt
@@ -2,7 +2,7 @@ package com.airongomes.eventsapi.domain.usecase
 
 import com.airongomes.eventsapi.domain.model.CheckIn
 import com.airongomes.eventsapi.domain.remote.NetworkResult
-import com.airongomes.eventsapi.domain.repository.EventRepository
+import com.airongomes.eventsapi.domain.repository.CheckInRepository
 import kotlinx.coroutines.flow.Flow
 
 interface CheckInUseCase {
@@ -10,7 +10,7 @@ interface CheckInUseCase {
 }
 
 class CheckInUseCaseImpl(
-    private val repository: EventRepository
+    private val repository: CheckInRepository
 ) : CheckInUseCase {
 
     override suspend fun invoke(checkIn: CheckIn) = repository.checkIn(checkIn)
