+++ app/src/test/java/com/airongomes/eventsapi/domain/repository/FakeCheckInRepository.kt
@@ -1,22 +1,20 @@
 package com.airongomes.eventsapi.domain.repository
 
-import com.airongomes.eventsapi.domain.model.Event
+import com.airongomes.eventsapi.domain.model.CheckIn
 import com.airongomes.eventsapi.domain.remote.NetworkResult
-import com.airongomes.eventsapi.fakeEvent
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
 
-class FakeEventRepository : EventRepository {
+class FakeCheckInRepository : CheckInRepository {
     var shouldReturnError: Boolean = false
 
-    override suspend fun getEventList(): Flow<NetworkResult<List<Event>>> =
+    override suspend fun checkIn(checkIn: CheckIn): Flow<NetworkResult<Any>> =
         if (shouldReturnError) flow { emit(NetworkResult.Error()) }
         else flow {
             emit(
-                NetworkResult.Success(
-                    data = listOf(fakeEvent)
-                )
+                NetworkResult.Success(data = Any())
             )
         }
+
 }
