+++ app/src/test/java/com/airongomes/eventsapi/viewModel/CheckInViewModelTest.kt
@@ -2,7 +2,7 @@ package com.airongomes.eventsapi.viewModel
 
 import com.airongomes.eventsapi.MainCoroutineRule
 import com.airongomes.eventsapi.domain.remote.NetworkResult
-import com.airongomes.eventsapi.domain.repository.FakeEventRepository
+import com.airongomes.eventsapi.domain.repository.FakeCheckInRepository
 import com.airongomes.eventsapi.domain.usecase.CheckInUseCase
 import com.airongomes.eventsapi.domain.usecase.CheckInUseCaseImpl
 import com.airongomes.eventsapi.fakeCheckIn
@@ -10,7 +10,8 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.drop
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.test.runTest
-import org.hamcrest.CoreMatchers.*
+import org.hamcrest.CoreMatchers.anything
+import org.hamcrest.CoreMatchers.instanceOf
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Before
 import org.junit.Rule
@@ -22,13 +23,13 @@ class CheckInViewModelTest {
     @get:Rule
     var mainCoroutineRule = MainCoroutineRule()
 
-    private lateinit var repository: FakeEventRepository
+    private lateinit var repository: FakeCheckInRepository
     private lateinit var checkInUseCase: CheckInUseCase
     private lateinit var viewModel: CheckInViewModel
 
     @Before
     fun setup() {
-        repository = FakeEventRepository()
+        repository = FakeCheckInRepository()
         checkInUseCase = CheckInUseCaseImpl(repository)
         viewModel = CheckInViewModel(checkInUseCase)
         viewModel.checkIn(fakeCheckIn)
